requirements.txt
asgiref==3.10.0
Django==5.1.14
django-taggit==6.1.0
sqlparse==0.5.3
tzdata==2025.2
